backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ML (must match your training versions)
scikit-learn==1.7.2
numpy==1.26.4
joblib==1.3.2
pandas==2.1.3

# Environment
python-dotenv==1.0.0